requirements.txt
numpy>=1.21.0
pandas>=1.3.0
torch>=1.9.0
yfinance>=0.1.63
pyyaml>=5.4.1
matplotlib>=3.4.0
scikit-learn>=0.24.0
pytest>=6.2.5
seaborn>=0.11.0 